// src/constants/enums/InGameItemsEnum.ts
/**
 * Enum bonding in-game item names to their item IDs.
 */
export const enum InGameItemsEnum {
    BLACK_DRAGON_LEATHER = '2509',
    BLACK_DRAGONHIDE_BODY = '2503',
    BLACK_DRAGONHIDE_CHAPS = '2497',
    BLACK_DRAGONHIDE_VAMBRACES = '2487',
    BLUE_DRAGON_LEATHER = '2505',
    BLUE_DRAGONHIDE_BODY = '2499',
    BLUE_DRAGONHIDE_CHAPS = '2493',
    BLUE_DRAGONHIDE_VAMBRACES = '2491',
    GREEN_DRAGON_LEATHER = '1745',
    GREEN_DRAGONHIDE_BODY = '1135',
    GREEN_DRAGONHIDE_CHAPS = '1099',
    GREEN_DRAGONHIDE_VAMBRACES = '1065',
    HAMMER = '2347',
    HARD_LEATHER = '1743',
    HARD_LEATHER_BODY = '1131',
    IRON_NAILS = '4820',
    LEATHER = '1741',
    LEATHER_GLOVES = '1059',
    LEATHER_BOOTS = '1061',
    LEATHER_COWL = '1167',
    LEATHER_VAMBRACES = '1063',
    LEATHER_BODY = '1129',
    LEATHER_CHAPS = '1095',
    NEEDLE = '1733',
    RED_DRAGON_LEATHER = '2507',
    RED_DRAGONHIDE_BODY = '2501',
    RED_DRAGONHIDE_CHAPS = '2495',
    RED_DRAGONHIDE_VAMBRACES = '2489',
    SNAKESKIN = '6289',
    SNAKESKIN_BOOTS = '6328',
    SNAKESKIN_VAMBRACES = '6330',
    SNAKESKIN_BANDANA = '6326',
    SNAKESKIN_CHAPS = '6324',
    SNAKESKIN_BODY = '6322',
    SNAKESKIN_SHIELD = '22272',
    THREAD = '1734',
    WILLOW_SHIELD = '22254',
}